add get datasets tree node public api.
